+++ lua/easy_en.lua
@@ -1,3 +1,6 @@
+local is_split_sentence
+local wordninja_split
+
 local function capture(cmd)
    local f = assert(io.popen(cmd, 'r'))
    local s = assert(f:read('*a'))
@@ -5,33 +8,55 @@ local function capture(cmd)
    return s
 end
 
-local function split_sentence_wordninja_py(sentence)
-   return capture([[python -c "import sys; import wordninja; sys.stdout.write(' '.join(wordninja.split(']] .. sentence .. [[')))"]])
-end
+local function init(env)
+   is_split_sentence = env.engine.schema.config:get_bool('easy_en/split_sentence')
+   if not is_split_sentence then
+      wordninja_split = function(sentence)
+         return sentence
+      end
+      return
+   end
 
-local function split_sentence_wordninja_rs(sentence, wordninja_rs_path)
-   return capture(wordninja_rs_path .. " -n " .. sentence)
-end
-
-local function split_sentence(sentence, env)
+   local use_wordninja_rs_lua_module = env.engine.schema.config:get_bool('easy_en/use_wordninja_rs_lua_module')
    local use_wordninja_rs = env.engine.schema.config:get_bool('easy_en/use_wordninja_rs')
+   local use_wordninja_py = env.engine.schema.config:get_bool('easy_en/use_wordninja_py')
+   if (not use_wordninja_rs_lua_module) and (not use_wordninja_rs) and (not use_wordninja_py) then
+      -- default use wordninja_rs_lua_module
+      use_wordninja_rs_lua_module = true
+   end
 
-   if (use_wordninja_rs) then
+   if use_wordninja_rs_lua_module then
+      local wordninja_rs_lua_module_path = env.engine.schema.config:get_string('easy_en/wordninja_rs_lua_module_path')
+      if not string.find(package.cpath, wordninja_rs_lua_module_path, 1, true) then
+         package.cpath = package.cpath .. ";" .. wordninja_rs_lua_module_path
+      end
+      wordninja_split = require("wordninja").split
+      return
+   end
+
+   if use_wordninja_rs then
       local wordninja_rs_path = env.engine.schema.config:get_string('easy_en/wordninja_rs_path')
-      return split_sentence_wordninja_rs(sentence, wordninja_rs_path)
-   else
-      return split_sentence_wordninja_py(sentence)
+      wordninja_split = function(sentence)
+         return capture(wordninja_rs_path .. " -n '" .. sentence .. "'")
+      end
+      return
+   end
+
+   if use_wordninja_py then
+      wordninja_split = function(sentence)
+         return capture([[python -c "import sys; import wordninja; sys.stdout.write(' '.join(wordninja.split(']] .. sentence .. [[')))"]])
+      end
+      return
    end
 end
 
 local function enhance_filter(input, env)
    local cands = {}
-   local is_split_sentence = env.engine.schema.config:get_bool('easy_en/split_sentence')
 
    for cand in input:iter() do
       if (cand.comment:find("☯")) then
          if (is_split_sentence) then
-            sentence = split_sentence(cand.text, env)
+            sentence = wordninja_split(cand.text)
             lower_sentence = string.lower(sentence)
 
             if (not (lower_sentence == sentence)) then
@@ -46,4 +71,4 @@ local function enhance_filter(input, env)
    end
 end
 
-return { enhance_filter = enhance_filter}
+return { enhance_filter = { init = init, func = enhance_filter} }