+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 0.9
+
+### new
+
+* 分词  
+  新的 lua 分词模块 [wordninja-rs-lua](https://github.com/BlindingDark/wordninja-rs-lua)  
+  在加载速度以及运行速度方面都有所提升  
+
+* 配置  
+  - 新增 `easy_en/use_wordninja_rs_lua_module` 配置
+  - 新增 `easy_en/use_wordninja_py` 配置
+  - 新增 `easy_en/wordninja_rs_lua_module_path` 配置，默认值为 `"/usr/lib/lua/5.4/wordninja.so"`
+
+### breaking change
+
+* 分词  
+  - 现在不再默认使用 `wordninja_rs`，而是使用 `wordninja_rs_lua_module`  
+
 ## 0.8.2
 
 ### new / fix